+++ src/Monad/Either.hs
@@ -9,13 +9,21 @@ module Monad.Either
        , leftToMaybe
        , rightToMaybe
        , maybeToEither
+       , whenLeft
+       , whenLeftM
+       , whenRight
+       , whenRightM
        ) where
 
-import           Data.Either   (Either (..), either, isLeft, isRight, lefts,
-                                partitionEithers, rights)
-import           Data.Function (const)
-import           Data.Maybe    (Maybe (..), maybe)
-import           Data.Monoid   (Monoid, mempty)
+import           Control.Applicative (Applicative)
+import           Control.Monad       (Monad (..))
+import           Data.Either         (Either (..), either, isLeft, isRight, lefts,
+                                      partitionEithers, rights)
+import           Data.Function       (const)
+import           Data.Maybe          (Maybe (..), maybe)
+import           Data.Monoid         (Monoid, mempty)
+
+import           Applicative         (pass)
 
 leftToMaybe :: Either l r -> Maybe l
 leftToMaybe = either Just (const Nothing)
@@ -32,3 +40,21 @@ maybeToLeft r = maybe (Right r) Left
 -- TODO: why this is @toEither@?
 maybeToEither :: Monoid b => (a -> b) -> Maybe a -> b
 maybeToEither = maybe mempty
+
+whenLeft :: Applicative f => Either l r -> (l -> f ()) -> f ()
+whenLeft (Left  l) f = f l
+whenLeft (Right _) _ = pass
+{-# INLINE whenLeft #-}
+
+whenLeftM :: Monad m => m (Either l r) -> (l -> m ()) -> m ()
+whenLeftM me f = me >>= \e -> whenLeft e f
+{-# INLINE whenLeftM #-}
+
+whenRight :: Applicative f => Either l r -> (r -> f ()) -> f ()
+whenRight (Left  _) _ = pass
+whenRight (Right r) f = f r
+{-# INLINE whenRight #-}
+
+whenRightM :: Monad m => m (Either l r) -> (r -> m ()) -> m ()
+whenRightM me f = me >>= \e -> whenRight e f
+{-# INLINE whenRightM #-}